adding, tests, linter, readme and light refactor
