+++ .babelrc
@@ -1,0 +1,12 @@
+{
+    "presets": [
+        "@babel/preset-env"
+    ],
+    "env": {
+        "test": {
+            "plugins": [
+                "@babel/plugin-transform-modules-commonjs"
+            ]
+        }
+    }
+}
