+++ index.html
@@ -40,7 +40,9 @@
 
     </div>
 
-    <script src="scripts/app.js"></script>
+    <script type="module" src="src/app.js"></script>
+    <script type="module" src="src/isSubset.mjs"></script>
+</body>
 </body>
 
 </html>
