+++ src/isSubset.mjs
@@ -1,0 +1,5 @@
+export default function isSubset(array1, array2) {
+  return array2.every((element) => {
+    return array1.includes(element);
+  });
+}
